Checkers big work!!
promotion does not work, its explained in the comments. peices can however not move backwards now!! also rand loops so its better.

+++ checkers.py
@@ -17,11 +17,28 @@ def clamp(input,min,max):
         return input
 
 def is_king(thing):
-    if thing == thing.upper:
+    if thing == thing.upper():
         return True
     else:
         return False
 
+def draw_board():
+    row_number = 0
+    print(top)
+    for row in board:
+        to_draw = ""
+        for space in row:
+            if space == "[0]":                            
+                to_draw = to_draw + f"[{bcolors.RED}{player_names[0][0]}{bcolors.WHITE}]"              #Where 0 and 1 become the colored player icons
+            elif space == "[1]":
+                to_draw = to_draw + f"[{bcolors.BLUE}{player_names[1][0]}{bcolors.WHITE}]"
+            else:                                                                           
+                to_draw = to_draw + space
+            
+        print(str(row_number+1)+" "+to_draw) #puts the numbers to left of board
+        row_number += 1
+
+    print(f"It is {player_names[who_move]}'s turn.")
 
 #  MAKE CHECKERS
 
@@ -103,26 +120,21 @@ who_not_move = 1
 
 #THIS IS THE GAME LOOP vvvvv
 
-while True:
-    row_number = 0
-    print(top)
-    for row in board:
-        to_draw = ""
-        for space in row:
-            if space == "[0]":                            
-                to_draw = to_draw + f"[{bcolors.RED}{player_names[0][0]}{bcolors.WHITE}]"              #Where 0 and 1 become the colored player icons
-            elif space == "[1]":
-                to_draw = to_draw + f"[{bcolors.BLUE}{player_names[1][0]}{bcolors.WHITE}]"
-            else:                                                                           
-                to_draw = to_draw + space
-            
-        print(str(row_number+1)+" "+to_draw) #puts the numbers to left of board
-        row_number += 1
+rand = False
+rand_tries = 0
 
-    print(f"It is {player_names[who_move]}'s turn.")
+while True:
+    #Theses two loops are important bc of stuff further down. The stuff for changing turn only should happen if the inner loop is broken (with break its not like.. broken broken.).
 
     while True:
-        game_input = input()
+
+        if (not rand) or rand_tries > 500:
+            rand_tries = 0
+            draw_board()
+            game_input = input()
+        else:
+            game_input = "rand"
+            rand_tries += 1
 
         if game_input == "stop":
             break
@@ -130,6 +142,8 @@ while True:
         #The inpupt structure for a move is naming the target spaces. IE a6b5, or a6c4 if you are capturing.
 
         if game_input == "rand": # this functionality is bad and terrible and a joke lmao
+            rand = True
+
             xy = [random.randint(1,8),random.randint(1,8)]
             plus_minus = [1,-1]
             plus_minus = plus_minus[random.randint(0,1)]
@@ -161,31 +175,51 @@ while True:
         # print(board[move_to_try[1]][move_to_try[0]][1])
 
         if board[move_to_try[1]][move_to_try[0]][1] == str(who_move): # Check if you own the peice #the one bracket at the end of the first expression [1] is for getting the middle symbol of [0] or [1] on the board. second character is [1].
+            
+            if (who_move == 0) != (move_to_try[1] > move_to_try[3]) or is_king(player_names[who_move][1]): #Are you trying to move backwards? (the != acts as an xor gate right now.) move_to_try[1] is the first y value, move_to_try[3] is where peice is going.
 
-            if abs(move_to_try[0]-move_to_try[2]) == 1 and abs(move_to_try[1]-move_to_try[3]) == 1: # Is this a standard move? (computer can't stop u from going backwards yet)
+                if abs(move_to_try[0]-move_to_try[2]) == 1 and abs(move_to_try[1]-move_to_try[3]) == 1: # Is this a regular (1 space) move? (computer can't stop u from going backwards yet)
 
-                if board[move_to_try[3]][move_to_try[2]] == "[ ]": # Is the target space empty?
+                    if board[move_to_try[3]][move_to_try[2]] == "[ ]": # Is the target space empty?
 
-                    board[move_to_try[1]][move_to_try[0]] = "[ ]" # remove your peice from where it was
-                    board[move_to_try[3]][move_to_try[2]] = "["+str(who_move)+"]" # place your peice where it goes
-                    break
-                
-            elif abs(move_to_try[0]-move_to_try[2]) == 2 and abs(move_to_try[1]-move_to_try[3]) == 2: # Is this a taking move?
-
-                if board[move_to_try[3]][move_to_try[2]] == "[ ]": #Is the target space empty?
-
-                    if board[int((move_to_try[3]+move_to_try[1])/2)][int((move_to_try[2]+move_to_try[0])/2)] == "["+str(who_not_move)+"]": # Is the space inbetween an enemy?
-
-                        board[int((move_to_try[3]+move_to_try[1])/2)][int((move_to_try[2]+move_to_try[0])/2)] = "[ ]" # kill enemy
+                        board[move_to_try[3]][move_to_try[2]] = board[move_to_try[1]][move_to_try[0]] # place your peice where it goes
                         board[move_to_try[1]][move_to_try[0]] = "[ ]" # remove your peice from where it was
-                        board[move_to_try[3]][move_to_try[2]] = "["+str(who_move)+"]" # place your peice where it goes
                         break
+                    
+                elif abs(move_to_try[0]-move_to_try[2]) == 2 and abs(move_to_try[1]-move_to_try[3]) == 2: # Is this a taking move (2 spaces)?
 
-                            #Need to change stuff with is_king()
+                    if board[move_to_try[3]][move_to_try[2]] == "[ ]": #Is the target space empty?
 
-            print("You can't move here!")
+                        if board[int((move_to_try[3]+move_to_try[1])/2)][int((move_to_try[2]+move_to_try[0])/2)] == "["+str(who_not_move)+"]": # Is the space inbetween an enemy?
+
+                            board[int((move_to_try[3]+move_to_try[1])/2)][int((move_to_try[2]+move_to_try[0])/2)] = "[ ]" # kill enemy
+                            board[move_to_try[3]][move_to_try[2]] = board[move_to_try[1]][move_to_try[0]] # place your peice where it goes
+                            board[move_to_try[1]][move_to_try[0]] = "[ ]" # remove your peice from where it was
+                            break
+
+                                #Need to change stuff with is_king()
+                else:
+                    print("You can't move here!")
+            else:
+                print("This peice can't move backwards!")
         else:
             print("This is not your peice!")
+
+#All this code is for if the move is successfull vvv the inner loop breaks and this happens when move works.
+
+
+
+#SUPER BIG ISSUE HERE: THE GAME'S DATA IS THE 1s and 0s, NOT LETTERS, THERFORE CAPITALISATION WILL NOT WORK. EITHER USE LIKE, 2 and 3 OR JUST MAKE LETTERS HOW THE GAME WORKS INTERPRETER IS GOOBER AHH SORTA
+
+
+    promo_y = board[move_to_try[3]] # This is for promoting peices
+
+    if (who_move == 0 and promo_y == 7) or (who_move == 1 and promo_y == 0): # This is for promoting peices
+        board[move_to_try[3]] = board[move_to_try[3]].upper() # This is for promoting peices
+
+
+
+    rand = False #If the move is successfull we stop trying to do random by changing this
 
     if who_move == 0: #change who's turn it is
         who_move = 1
